+++ packages/zowe-explorer/__tests__/__unit__/extension.unit.test.ts
@@ -119,6 +119,7 @@ async function createGlobalMocks() {
             disableValidationContext: jest.fn(),
             enableValidationContext: jest.fn(),
             validationArraySetup: jest.fn(),
+            fetchAllProfiles: jest.fn(),
             getProfileInfo: () => createInstanceOfProfileInfo(),
         },
         mockExtension: null,
