+++ packages/zowe-explorer/src/shared/refresh.ts
@@ -12,7 +12,7 @@
 import { IZoweTree, IZoweTreeNode, PersistenceSchemaEnum } from "@zowe/zowe-explorer-api";
 import { PersistentFilters } from "../PersistentFilters";
 import { Profiles } from "../Profiles";
-import { syncSessionNode } from "../utils/ProfilesUtils";
+import { removeSession, syncSessionNode } from "../utils/ProfilesUtils";
 import { ZoweExplorerApiRegister } from "../ZoweExplorerApiRegister";
 import { resetValidationSettings, returnIconState } from "./actions";
 import { labelRefresh } from "./utils";
@@ -52,56 +52,4 @@ export async function refreshAll(treeProvider: IZoweTree<IZoweTreeNode>) {
             await removeSession(treeProvider, sessNode.label.toString().trim());
         }
     });
-}
-
-export async function removeSession(treeProvider: IZoweTree<IZoweTreeNode>, profileName: string): Promise<void> {
-    const treeType = treeProvider.getTreeType();
-    let schema;
-    switch (treeType) {
-        case PersistenceSchemaEnum.Dataset:
-            schema = globals.SETTINGS_DS_HISTORY;
-            break;
-        case PersistenceSchemaEnum.USS:
-            schema = globals.SETTINGS_USS_HISTORY;
-            break;
-        case PersistenceSchemaEnum.Job:
-            schema = globals.SETTINGS_JOBS_HISTORY;
-            break;
-    }
-    if (treeType !== globals.SETTINGS_JOBS_HISTORY) {
-        // Delete from file history
-        const fileHistory: string[] = treeProvider.getFileHistory();
-        fileHistory
-            .slice()
-            .reverse()
-            .filter((item) => item.substring(1, item.indexOf("]")).trim() === profileName.toUpperCase())
-            .forEach((file) => {
-                treeProvider.removeFileHistory(file);
-            });
-    }
-    // Delete from Favorites
-    treeProvider.removeFavProfile(profileName, false);
-    // Delete from Tree
-    treeProvider.mSessionNodes.forEach((sessNode) => {
-        if (sessNode.getProfileName() === profileName) {
-            treeProvider.deleteSession(sessNode);
-            sessNode.dirty = true;
-            treeProvider.refresh();
-        }
-    });
-    // Delete from Sessions list
-    const setting: any = {
-        ...vscode.workspace.getConfiguration().get(schema),
-    };
-    let sess: string[] = setting.sessions;
-    let fave: string[] = setting.favorites;
-    sess = sess.filter((value) => {
-        return value.trim() !== profileName;
-    });
-    fave = fave.filter((element) => {
-        return element.substring(1, element.indexOf("]")).trim() !== profileName;
-    });
-    setting.sessions = sess;
-    setting.favorites = fave;
-    await vscode.workspace.getConfiguration().update(schema, setting, vscode.ConfigurationTarget.Global);
 }
